rearranged code and finished ThankYouPage ui
