+++ client/src/components/Login/login.js
@@ -1,12 +1,42 @@
 import React from 'react'
-import LoginPage from '../../pages/Login'
+import PosmeLogo from './images/POSME.png'
+import './login.css'
+import TextField from '@material-ui/core/TextField'
 
-class Login extends React.Component {
-  render () {
-    return (
-      <LoginPage />
-    )
-  }
+const Login = _ => {
+  return (
+    <>
+      <img alt='logo' className='posme' src={PosmeLogo} />
+      <h4 className='meaning'>Point Of Sale Made Easy</h4>
+      <form className='form-input' >
+        <TextField
+          label='Username'
+          className='username-input'
+          type='username'
+          autoComplete='username'
+          margin='normal'
+          variant='outlined'
+        />
+        <TextField
+          label='Password'
+          className='username-input'
+          type='password'
+          autoComplete='current-password'
+          margin='normal'
+          variant='outlined'
+        />
+        <div>
+          <button
+            className='login'
+            type='button'
+            name='data' >
+            LOGIN
+          </button>
+        </div>
+      </form>
+    </>
+  )
 }
 
 export default Login
+
