+++ client/src/components/ThankYou/images/giphy.png
