+++ client/src/components/ThankYou/index.js
@@ -1,0 +1,1 @@
+export { default } from './ThankYou.js'
